Added card view implementations, added skeleton loaders, improved pagination text
