+++ package.json
@@ -77,6 +77,7 @@
                 "luxon": "^3.5.0",
                 "maplibre-gl": "^5.1.0",
                 "next": "15.1.6",
+                "nuqs": "^2.4.3",
                 "react": "19.0.0",
                 "react-chartjs-2": "^5.3.0",
                 "react-dom": "19.0.0",
