--- src/app/_search-new/_components/result.module.css
@@ -1,12 +1,0 @@
-.result {
-  border-radius: var(--mantine-radius-sm);
-
-  &:first-of-type {
-    border-top-left-radius: var(--mantine-radius-md);
-    border-top-right-radius: var(--mantine-radius-md);
-  }
-  &:last-of-type {
-    border-bottom-left-radius: var(--mantine-radius-md);
-    border-bottom-right-radius: var(--mantine-radius-md);
-  }
-}
