+++ src/app/search-new/_components/result.module.css
@@ -1,0 +1,24 @@
+.result {
+  border-radius: var(--mantine-radius-sm);
+  transition: opacity ease 200ms;
+  &:hover {
+    opacity: 0.4;
+  }
+
+  &:first-of-type {
+    border-top-left-radius: var(--mantine-radius-lg);
+    border-top-right-radius: var(--mantine-radius-lg);
+  }
+  &:last-of-type {
+    border-bottom-left-radius: var(--mantine-radius-lg);
+    border-bottom-right-radius: var(--mantine-radius-lg);
+  }
+}
+
+.card {
+  transition: all ease 200ms;
+  &:hover {
+    transform: scale(1.05);
+    opacity: 0.6;
+  }
+}
